requirements.txt
fastapi
uvicorn
requests
beautifulsoup4
schedule
python-dotenv
a2wsgi
paramiko
